// frontend/src/lib/api/models/QuizAvailabilityResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * GET availability wire shape (Stage 5 HTTP contract).
 */
export type QuizAvailabilityResponse = {
    availability: string;
    reasonCode?: (string | null);
};